legacy/requirements.txt
streamlit
requests
psycopg2-binary
python-dotenv
sqlalchemy
pandas
